[feature/Inhabas#228] 공모전 게시판 API 구현

+++ resource-server/src/main/java/com/inhabas/api/domain/contest/repository/ContestBoardRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.inhabas.api.domain.contest.domain.ContestBoard;
 
 public interface ContestBoardRepository
-    extends JpaRepository<ContestBoard, Integer>, ContestBoardRepositoryCustom {}
+    extends JpaRepository<ContestBoard, Long>, ContestBoardRepositoryCustom {}
